feat: validateAssistant
